requirements.txt
# QRes OS 4 - Restaurant Management System
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
aiosqlite==0.19.0
asyncpg==0.29.0  # For PostgreSQL support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
httpx==0.25.2
qrcode[pil]==7.4.2
python-slugify==8.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
